hospital/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
